(build) Mini overhaul of KTS file
I initially started updating this KTS file thinking that I was going to
be adding the SonarQube analysis build to it, however, I then realised
that this repository is not using Chocolatey.Cake.Recipe, which means
there is no easy way to add SonarQube to this repository.  Not
without duplicating a lot of the logic from the recipe package into the
build script in this repository.